update DeepSEA/beluga to pytorch model
