+++ DeepSEA/beluga/model.py
@@ -1,29 +1,54 @@
-from kipoi.model import BaseModel
-import torch
+from torch import nn
 
 
-class TorchModel(BaseModel):
+class LambdaBase(nn.Sequential):
+    def __init__(self, fn, *args):
+        super(LambdaBase, self).__init__(*args)
+        self.lambda_func = fn
 
-    def __init__(self, torch_t7_file=None, auto_use_cuda=True):
+    def forward_prepare(self, input):
+        output = []
+        for module in self._modules.values():
+            output.append(module(input))
+        return output if output else input
 
-        from torch.utils.serialization import load_lua
-        self.model = load_lua(torch_t7_file)
-        self.model.evaluate()
 
-        if auto_use_cuda and torch.cuda.is_available():
-            self.model = self.model.cuda()
-            self.use_cuda = True
-        else:
-            self.use_cuda = False
+class Lambda(LambdaBase):
+    def forward(self, input):
+        return self.lambda_func(self.forward_prepare(input))
 
-    def predict_on_batch(self, x):
-        # numpy -> torch
-        x_torch = torch.from_numpy(x)
-        if self.use_cuda:
-            x_torch = x_torch.cuda()
-        preds_torch = self.model.forward(x_torch)
 
-        # torch -> numpy
-        if self.use_cuda:
-            preds_torch = preds_torch.cpu()
-        return preds_torch.numpy().copy()
+class Beluga(nn.Module):
+    def __init__(self):
+        super(Beluga, self).__init__()
+        self.model = nn.Sequential(
+            nn.Sequential(
+                nn.Conv2d(4, 320, (1, 8)),
+                nn.ReLU(),
+                nn.Conv2d(320, 320, (1, 8)),
+                nn.ReLU(),
+                nn.Dropout(0.2),
+                nn.MaxPool2d((1, 4), (1, 4)),
+                nn.Conv2d(320, 480, (1, 8)),
+                nn.ReLU(),
+                nn.Conv2d(480, 480, (1, 8)),
+                nn.ReLU(),
+                nn.Dropout(0.2),
+                nn.MaxPool2d((1, 4), (1, 4)),
+                nn.Conv2d(480, 640, (1, 8)),
+                nn.ReLU(),
+                nn.Conv2d(640, 640, (1, 8)),
+                nn.ReLU(),
+            ),
+            nn.Sequential(
+                nn.Dropout(0.5),
+                Lambda(lambda x: x.view(x.size(0), -1)),
+                nn.Sequential(Lambda(lambda x: x.view(1, -1) if 1 == len(x.size()) else x), nn.Linear(67840, 2003)),
+                nn.ReLU(),
+                nn.Sequential(Lambda(lambda x: x.view(1, -1) if 1 == len(x.size()) else x), nn.Linear(2003, 2002)),
+            ),
+            nn.Sigmoid(),
+        )
+
+    def forward(self, x):
+        return self.model(x)
